Remove duplicate EA HBM subtree, rename membw to membw_analysis, add design comments
- Remove ea_hbm_bw_bound subtree (identical formulas to gl2_mem_bw_bound)
- Remove ea_hbm_read/write/other guidance templates
- Rename src/membw -> src/membw_analysis, tests/unit/membw -> tests/unit/membw_analysis
- Add membw_analysis_dir() path helper, consolidate _guidance_path into caller
- Add parent gating design-choice comment in engine.py
- Add schema hash contract comment in tree_spec.py
- Update pre-commit hook file filter and validation script imports
