+++ projects/rocprofiler-compute/.pre-commit-config.yaml
@@ -77,7 +77,7 @@ repos:
         name: Membw tree spec metric validation
         entry: bash -lc 'cd projects/rocprofiler-compute && python3 tools/validate_membw_tree_spec.py'
         language: system
-        files: (_membw_tree_spec\.yaml|_membw_guidance\.yaml|3000_mem_bw.*\.yaml|membw/models\.py)$
+        files: (_membw_tree_spec\.yaml|_membw_guidance\.yaml|3000_mem_bw.*\.yaml|membw_analysis/models\.py)$
         pass_filenames: false
         stages: [pre-commit]
